requirements.txt
django ~= 4.0
django-extensions ~= 3.1
django-cors-headers ~= 3.10

Pillow ~= 8.4
psycopg2-binary ~= 2.9
gunicorn ~= 20.1

djangorestframework ~= 3.13
djangorestframework-simplejwt ~= 5.0
drf-access-policy ~= 1.1.1

transbank-sdk ~= 2.0

reportlab == 3.6.6 ## https://github.com/xhtml2pdf/xhtml2pdf/issues/589
xhtml2pdf ~= 0.2
pyqrcode ~= 1.2
pypng ~= 0.0
PyPDF2 ~= 3.0

google-api-python-client ~= 2.33
google-auth-httplib2 ~= 0.1
google-auth-oauthlib ~= 0.4

unidecode ~= 1.3
babel ~= 2.9

drf-spectacular ~= 0.22